Reformatted, removed todos, boilerplate and unused code.
