+++ src/main/java/com/sda/eventapp/repository/UserRepository.java
@@ -4,5 +4,4 @@ import com.sda.eventapp.model.User;
 import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Long> {
-    User findById(long id);
 }
