--- src/main/java/com/sda/eventapp/service/CommentService.java
@@ -1,16 +1,0 @@
-package com.sda.eventapp.service;
-
-import com.sda.eventapp.repository.CommentRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Service
-@Slf4j
-@RequiredArgsConstructor
-public class CommentService {
-
-    private final CommentRepository commentRepository;
-
-
-}
