+++ src/main/java/com/sda/eventapp/web/dto/EventWithBasicData.java
@@ -1,22 +1,17 @@
 package com.sda.eventapp.web.dto;
 
-
-
-import lombok.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.time.LocalDateTime;
-
 
 @Getter
 @Setter
 @Builder
 public class EventWithBasicData {
-
     private String title;
-
     private String description;
-
     private LocalDateTime startingDateTime;
-
     private LocalDateTime endingDateTime;
 }
